refactoring actions complete - fix the liveUpdate!
